feat: 로또 번호 오름차순 출력 기능

+++ src/main/java/lotto/utils/Calculator.java
@@ -64,7 +64,7 @@ public class Calculator {
 
     public static double calculateEarningRate(int amount, List<Integer> ranks){
         int totalProfit = calculateTotalProfit(ranks);
-        double earningRate = (totalProfit / amount)*100;
+        double earningRate = ((double)totalProfit / (double)amount)*100;
         return earningRate;
     }
